+++ init-submodules.sh
@@ -5,9 +5,4 @@
 #
 # You do not need to call this script if you only intend to build bare-metal workloads.
 
-git submodule update --init \
-  riscv-linux \
-  riscv-pk \
-  wlutil/busybox \
-  wlutil/br/buildroot \
-  boards/firechip/drivers/*
+git submodule update --init